Move all code to /src
