+++ src/commands/forecast.js
